requirements.txt
aiogram>=3.0.0
python-dotenv>=0.19.0
aiohttp>=3.8.0
Pillow>=9.0.0
numpy>=1.21.0
pytest>=7.0.0
pytest-asyncio>=0.18.0
pytest-cov>=3.0.0
black==23.12.1
isort==5.13.2
flake8==7.0.0
mypy==1.8.0
beautifulsoup4>=4.9.3
selenium>=4.0.0
webdriver-manager>=3.8.0